Pulled in fabric's new head and then switched back to Process.
I'll wait to try and implement my own Pool, as the lack of shared state is
killing the fabric tasks. They lose all knowledge of their env and the specific
host they are on.

I need to just make pool type class/function that just makes a bubble of size x
of Process()es to run, and keeps it to a said size until it completes.