It works!
So I have added in the Job_Queue class, and incorperated it into the main fab
section to manage the multiprocessing Processes.

I have hooked into the state.output to toggle the _debug flag for job_queue.

The job queue uses the host_string var that gets set as the Process().name to
then set the env.state.host_string again inside the running queue.